refactor: move service types under types
