+++ packages/core/src/services/ArtifactsService.ts
@@ -1,10 +1,7 @@
-import { Artifact } from "../types/hardhat";
-import { Providers } from "../types/providers";
+import type { Artifact } from "../types/hardhat";
+import type { Providers } from "../types/providers";
 
-export interface IArtifactsService {
-  getArtifact(name: string): Promise<Artifact>;
-  hasArtifact(name: string): Promise<boolean>;
-}
+import { IArtifactsService } from "./services";
 
 export class ArtifactsService implements IArtifactsService {
   constructor(private readonly _providers: Providers) {}
