+++ packages/core/src/services/ConfigService.ts
@@ -1,11 +1,7 @@
-import { ExternalParamValue } from "../types/deploymentGraph";
-import { HasParamResult, Providers } from "../types/providers";
+import type { ExternalParamValue } from "../types/deploymentGraph";
+import type { HasParamResult, Providers } from "../types/providers";
 
-export interface IConfigService {
-  getParam(paramName: string): Promise<ExternalParamValue>;
-
-  hasParam(paramName: string): Promise<HasParamResult>;
-}
+import { IConfigService } from "./services";
 
 export class ConfigService implements IConfigService {
   constructor(private readonly _providers: Providers) {}
