+++ packages/core/src/services/NetworkService.ts
@@ -1,8 +1,6 @@
 import { Providers } from "../types/providers";
 
-export interface INetworkService {
-  getChainId(): Promise<number>;
-}
+import { INetworkService } from "./services";
 
 export class NetworkService implements INetworkService {
   constructor(private readonly _providers: Providers) {}
