+++ packages/core/src/validation/dispatch/helpers.ts
@@ -1,4 +1,4 @@
-import type { Services } from "../../services/types";
+import type { Services } from "../../services/services";
 import type { CallableFuture } from "../../types/future";
 
 import {
