+++ packages/core/test/helpers.ts
@@ -1,12 +1,15 @@
 import { ethers } from "ethers";
 
-import { IAccountsService } from "../src/services/AccountsService";
-import { IArtifactsService } from "../src/services/ArtifactsService";
-import { IConfigService } from "../src/services/ConfigService";
-import { IContractsService } from "../src/services/ContractsService";
-import { INetworkService } from "../src/services/NetworkService";
-import { ITransactionsService } from "../src/services/TransactionsService";
-import { Services, TransactionOptions } from "../src/services/types";
+import {
+  Services,
+  TransactionOptions,
+  ITransactionsService,
+  INetworkService,
+  IContractsService,
+  IConfigService,
+  IArtifactsService,
+  IAccountsService,
+} from "../src/services/services";
 import { Artifact } from "../src/types/hardhat";
 import { HasParamResult } from "../src/types/providers";
 import { IgnitionError } from "../src/utils/errors";
