+++ packages/core/test/validation.ts
@@ -7,7 +7,7 @@ import sinon from "sinon";
 
 import { buildModule } from "../src/dsl/buildModule";
 import { generateDeploymentGraphFrom } from "../src/process/generateDeploymentGraphFrom";
-import { Services } from "../src/services/types";
+import { Services } from "../src/services/services";
 import { ArtifactContract } from "../src/types/future";
 import { Artifact } from "../src/types/hardhat";
 import { Module, ModuleDict } from "../src/types/module";
